+++ src/ts/index.ts
@@ -175,6 +175,7 @@ import MantineProvider from './components/styles/MantineProvider';
 import ChipGroup from './components/core/chip/ChipGroup';
 import RichTextEditor from './components/extensions/richtexteditor/RichTextEditor';
 import TypographyStylesProvider from './components/styles/TypographyStylesProvider';
+import DirectionProvider from './components/styles/DirectionProvider';
 
 export {
     Accordion,
@@ -353,4 +354,5 @@ export {
     RichTextEditor,
     TypographyStylesProvider,
     appNotifications,
+    DirectionProvider,
 };