+++ CHANGELOG.md
@@ -4,12 +4,17 @@
 
 
 ### Added
-- New `scrollTo` prop for `ScrollArea` #645
-- New `ScrollAreaAutoheight` component.  #645 by @AnnMarieW
+- New `scrollTo` prop for `ScrollArea` #645 by @AnnMarieW
+- New `ScrollAreaAutoheight` component. #645 
+
+- added `DirectionProvider` to handle RTL (right-to-left) text direction.  #650 by @AnnMarieW
 
 ### Fixed
 
 - BarChart: Added default `valueFormatter` to prevent rendering issues when `valueFormatter` is undefined. #464 by @AnnMarieW
+
+### Changed
+- Upgraded to latest Mantine (8.3.1) #650 by @AnnMarieW
 
 
 # 2.2.1